refactor(coloc): colocalisation step cleanup
