+++ docs/python_api/datasets/l2g_features/colocalisation.md
@@ -22,3 +22,4 @@ title: From colocalisation
 ::: gentropy.dataset.l2g_features.colocalisation.common_colocalisation_feature_logic
 ::: gentropy.dataset.l2g_features.colocalisation.extend_missing_colocalisation_to_neighbourhood_genes
 ::: gentropy.dataset.l2g_features.colocalisation.common_neighbourhood_colocalisation_feature_logic
+::: gentropy.dataset.l2g_features.colocalisation.extract_maximum_coloc_probability_per_region_and_gene
