+++ docs/python_api/methods/joint_colocalisation.md
@@ -1,0 +1,5 @@
+---
+title: ColocPIP Ecaviar
+---
+
+::: gentropy.method.colocalisation.ColocPIPECaviar
