+++ tests/gentropy/dataset/test_colocalisation.py
@@ -117,24 +117,6 @@ class TestAppendStudyMetadata:
         )
 
 
-def test_extract_maximum_coloc_probability_per_region_and_gene(
-    mock_colocalisation: Colocalisation,
-    mock_study_locus: StudyLocus,
-    mock_study_index: StudyIndex,
-    filter_by_colocalisation_method: str | None = None,
-) -> None:
-    """Test extracting maximum coloc probability per region and gene returns a dataframe with the correct columns: studyLocusId, geneId, h4."""
-    filter_by_colocalisation_method = filter_by_colocalisation_method or "Coloc"
-    res_df = mock_colocalisation.extract_maximum_coloc_probability_per_region_and_gene(
-        mock_study_locus,
-        mock_study_index,
-        filter_by_colocalisation_method=filter_by_colocalisation_method,
-    )
-    expected_cols = ["studyLocusId", "geneId", "h4"]
-    for col in expected_cols:
-        assert col in res_df.columns, f"Column {col} not found in result DataFrame."
-
-
 def test_drop_trans_effects(spark: SparkSession) -> None:
     """Test filtering out trans effects from QTLs."""
     sample_study_locus = StudyLocus(
