+++ tests/gentropy/dataset/test_l2g_feature.py
@@ -39,6 +39,7 @@ from gentropy.dataset.l2g_features.colocalisation import (
     common_colocalisation_feature_logic,
     common_neighbourhood_colocalisation_feature_logic,
     extend_missing_colocalisation_to_neighbourhood_genes,
+    extract_maximum_coloc_probability_per_region_and_gene,
 )
 from gentropy.dataset.l2g_features.distance import (
     DistanceFootprintMeanFeature,
@@ -80,6 +81,25 @@ from gentropy.method.l2g.feature_factory import L2GFeatureInputLoader
 
 if TYPE_CHECKING:
     from pyspark.sql import SparkSession
+
+
+def test_extract_maximum_coloc_probability_per_region_and_gene(
+    mock_colocalisation: Colocalisation,
+    mock_study_locus: StudyLocus,
+    mock_study_index: StudyIndex,
+    filter_by_colocalisation_method: str | None = None,
+) -> None:
+    """Test extracting maximum coloc probability per region and gene returns a dataframe with the correct columns: studyLocusId, geneId, h4."""
+    filter_by_colocalisation_method = filter_by_colocalisation_method or "Coloc"
+    res_df = extract_maximum_coloc_probability_per_region_and_gene(
+        mock_colocalisation,
+        mock_study_locus,
+        mock_study_index,
+        filter_by_colocalisation_method=filter_by_colocalisation_method,
+    )
+    expected_cols = ["studyLocusId", "geneId", "h4"]
+    for col in expected_cols:
+        assert col in res_df.columns, f"Column {col} not found in result DataFrame."
 
 
 @pytest.mark.parametrize(
@@ -383,7 +403,7 @@ class TestCommonColocalisationFeatureLogic:
         """Set up the test variables."""
         self.colocalisation_method = "Coloc"
         self.colocalisation_metric = "h4"
-        self.qtl_type = "eqtl"
+        self.qtl_type = ["eqtl"]
 
         self.sample_study_loci_to_annotate = StudyLocus(
             _df=spark.createDataFrame(
